% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotate_gRNA}
\alias{annotate_gRNA}
\title{Creates a metadata column in a Seurat object based on expression of a list of features, labeling each cell with its barcode, "Neg" or "Multiple".}
\usage{
annotate_gRNA(obj, gRNA_list)
}
\arguments{
\item{obj}{Seurat v5 object}

\item{gRNA_list}{A list of features or genes for testing presence/absence of}
}
\value{
Seurat object with new metadata column "gRNA" containing entries "Neg", "Multiple" or name of the gRNA
}
\description{
Creates a metadata column in a Seurat object based on expression of a list of features, labeling each cell with its barcode, "Neg" or "Multiple".
}
\author{
Kimberle Shen
}
