% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotate_pos_neg}
\alias{annotate_pos_neg}
\title{Creates a metadata column based on expression of a user-given list of barcodes, with values "Pos" or "Neg".}
\usage{
annotate_pos_neg(obj, gRNA_list)
}
\arguments{
\item{obj}{Seurat v5 object with expression of features in gRNA_list (barcodes)}

\item{gRNA_list}{A list of features or genes for testing the presence/absence of}
}
\value{
Seurat object with new metadata column "gRNA_exp" containing entries "Pos" or "Neg"
}
\description{
Creates a metadata column based on expression of a user-given list of barcodes, with values "Pos" or "Neg".
}
\author{
Kimberle Shen
}
