% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{barplot_barcode_count}
\alias{barplot_barcode_count}
\title{Bar plot showing the number of cells for each perturbation.}
\usage{
barplot_barcode_count(obj, Multiple = FALSE, Neg = FALSE)
}
\arguments{
\item{obj}{Seurat v5 object with annotate_gRNA ran.}

\item{Multiple}{Boolean. If FALSE, do not show "Multiple" cells.}

\item{Neg}{Boolean. If FALSE, do not show "Neg" cells.}
}
\value{
ggplot2 barplot
}
\description{
Bar plot showing the number of cells for each perturbation.
}
\author{
Kimberle Shen
}
