% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{barplot_pos}
\alias{barplot_pos}
\title{Creates barplot showing the \% of barcode-positive cells in each identity class. Ident can be seurat cluster, or annotated cell type, etc.}
\usage{
barplot_pos(obj, ident = seurat_clusters)
}
\arguments{
\item{obj}{Seurat v5 object with annotate_pos_neg function ran}

\item{ident}{Idents to plot. Default is seurat_clusters, which requres seurat pipeline to be ran on obj beforehand.}
}
\value{
Bar plot showing the percentage of positive cells for each Ident
}
\description{
Creates barplot showing the \% of barcode-positive cells in each identity class. Ident can be seurat cluster, or annotated cell type, etc.
}
\author{
Kimberle Shen
}
