% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_euclidean_neighbours}
\alias{get_euclidean_neighbours}
\title{Given a list of cellIDs and the Seurat object, returns the cellIDs of its nearest neighbours (euclidean distance)}
\usage{
get_euclidean_neighbours(obj, cell_list, k_geom = 15)
}
\arguments{
\item{obj}{Seurat v5 object}

\item{cell_list}{list of cellIDs in char format, eg c("65828963750382", "65816078848465", "65584150614472")}

\item{k_geom}{number of neighbours for each cell, default is 15}
}
\value{
list of cellIDs in char format
}
\description{
Adapted from: https://cran.rstudio.com/web/packages/RcppHNSW/index.html
}
\author{
Kimberle Shen
}
