% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_KNN}
\alias{get_KNN}
\title{Given a list of CellIDs, return CellIDs of K nearest neighbours.}
\usage{
get_KNN(obj, cell_list, k_geom = 15)
}
\arguments{
\item{obj}{Seurat v5 object}

\item{cell_list}{list of cellIDs in char format, eg c("65828963750382", "65816078848465", "65584150614472")}

\item{k_geom}{number of neighbours for each cell, default is 15}
}
\value{
list of cellIDs in char format
}
\description{
Adapted from: https://github.com/prabhakarlab/Banksy
}
\author{
Kimberle Shen
}
