% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_radial_neighbours}
\alias{get_radial_neighbours}
\title{Given a list of CellIDs, return CellIDs of neighbours within a certain radius.}
\usage{
get_radial_neighbours(obj, cell_list, radius = 200)
}
\arguments{
\item{obj}{Seurat v5 object}

\item{cell_list}{list of cellIDs in char format, eg c("65828963750382", "65816078848465", "65584150614472")}

\item{radius}{radius for each to identify neighbourgood of a cell, default is 200}
}
\value{
list of list (neighbour cellIDs for each input cellID)
}
\description{
Adapted from: https://cran.rstudio.com/web/packages/RcppHNSW/index.html
}
\author{
Kimberle Shen
}
