% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{liana_table}
\alias{liana_table}
\title{Interrogates the LIANA database to identify LR interactions, visualizating cell-cell communication between 2 groups of cells (source group and target group)}
\usage{
liana_table(obj, control = "Neg")
}
\arguments{
\item{obj}{Seurat v5 object with "data" layer and annotateWhichgRNA function run to produce a gRNA metadata column}

\item{control}{control cells to compare with, for the calculation of DEGs, default is "Neg", but should ideally be a proper control like "mSafe".}
}
\value{
Seurat object with new metadata column "gRNA_exp" containing entries "Pos" or "Neg"
}
\description{
Interrogates the LIANA database to identify LR interactions, visualizating cell-cell communication between 2 groups of cells (source group and target group)
}
\author{
Kimberle Shen
}
