% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{module_score}
\alias{module_score}
\title{VlnPlot showing module score of each perturbation, with statistics calculated and shown.}
\usage{
module_score(obj, module, idents = unique(Idents(obj)), increment = 0.25)
}
\arguments{
\item{obj}{Seurat v5 object with AddModuleScore ran for at least 1 module}

\item{module}{The module score to be plotted, eg "Myelination_score1"}

\item{idents}{A list of identities of the Seurat object to plot and calculate. List needs 2 or more elements.}

\item{increment}{A number that adjusts the spacing of the significance bars. Default is 0.25 (25\%)}
}
\value{
ggplot2 VlnPlot
}
\description{
VlnPlot showing module score of each perturbation, with statistics calculated and shown.
}
\author{
Kimberle Shen
}
