% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nndist_pos_neg}
\alias{nndist_pos_neg}
\title{Frequency histogram of Nearest Neighbor distances for Pos points in a spatial region, highlighting distribution of distances.}
\usage{
nndist_pos_neg(obj)
}
\arguments{
\item{obj}{Seurat v5 object with annotate_pos_neg function ran}
}
\value{
Histogram plot of neaest neighbour distances for positive points.
}
\description{
A left-skewed graph suggests clustering of Pos points while a right-skewed graph suggests a more dispersed pattern. A uniform graph could indicate random distribution without clustering or dispersion, while multimodal peaks could indicate multiple clusters.
}
\author{
Kimberle Shen
}
