% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{piechart_barcode_uniqueness}
\alias{piechart_barcode_uniqueness}
\title{Creates a piechart showing the proportion of cells having 1, 2, 3 etc unique features. In most cases, most cells ideally have only 1 unique barcode}
\usage{
piechart_barcode_uniqueness(obj, gRNA_list, remove0 = TRUE)
}
\arguments{
\item{obj}{Seurat v5 object with annotate_gRNA ran}

\item{gRNA_list}{A list of features or genes for testing presence/absence of}

\item{remove0}{Boolean value whether to remove cells containing 0 features or not. Default is TRUE.}
}
\value{
ggplot2 piechart
}
\description{
Creates a piechart showing the proportion of cells having 1, 2, 3 etc unique features. In most cases, most cells ideally have only 1 unique barcode
}
\author{
Kimberle Shen
}
