% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_barcode_spread}
\alias{plot_barcode_spread}
\title{Creates scatterplot showing spatial location of different perturbations in the tissue}
\usage{
plot_barcode_spread(obj, pt.size = 1)
}
\arguments{
\item{obj}{Seurat v5 object with annotate_gRNA ran}

\item{pt.size}{Size of points on the scatter plot}

\item{ident}{metadata column where gRNA names are stored}
}
\value{
ggplot2 scatterplot
}
\description{
Creates scatterplot showing spatial location of different perturbations in the tissue
}
\author{
Kimberle Shen
}
