% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_ligand_receptor}
\alias{plot_ligand_receptor}
\title{Plot of each cell and their neighbours, with each cell coloured by expression level of a specified ligand/receptor pair.}
\usage{
plot_ligand_receptor(obj, cell_list_source, cell_list_target, ligand, receptor)
}
\arguments{
\item{obj}{Seurat v5 object with annotate_gRNA function ran}

\item{cell_list_source}{List of cell names to colour the ligand's expression (red)}

\item{cell_list_target}{List of cell names (neighbours of cell_list_source) to colour the receptor's expression (blue)}

\item{ligand}{Name of ligand (gene name)}

\item{receptor}{Name of receptor (gene name)}
}
\value{
Seurat object with spatial coordinates
}
\description{
Plot of each cell and their neighbours, with each cell coloured by expression level of a specified ligand/receptor pair.
}
\author{
Kimberle Shen
}
