% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_perturbation_effect}
\alias{plot_perturbation_effect}
\title{Creates 2 barplots: number of DEGs and average fold change of DEGs, for each perturbation}
\usage{
plot_perturbation_effect(
  obj,
  perturbations,
  control = "Neg",
  logfc.threshold = 0.25,
  min.pct = 0.2
)
}
\arguments{
\item{obj}{Seurat v5 object with annotate_gRNA ran}

\item{perturbations}{Identity names of perturbations to test for}

\item{control}{control cells to compare with, for the calculation of DEGs, default is "Neg", but should ideally be a proper control like "mSafe".}

\item{logfc.threshold}{Log2FC cutoff for counting of DEGs, default is 0.25}

\item{min.pct}{Only consider genes with expressed in this minimum percentage of cells. Default 0.2.}
}
\value{
ggplot2 barplots
}
\description{
Creates 2 barplots: number of DEGs and average fold change of DEGs, for each perturbation
}
\author{
Kimberle Shen
}
