% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pos_neg_DEG}
\alias{pos_neg_DEG}
\title{Calculates DEGs for Pos vs Neg cell populations and shows top upregulated genes for each population.}
\usage{
pos_neg_DEG(obj, min.pct = 0.2, logfc.threshold = 0.25, n = 10)
}
\arguments{
\item{obj}{Seurat v5 object with annotate_pos_neg function ran}

\item{min.pct}{Only consider genes expressed in this minimum percentage of cells. Default 0.2.}

\item{logfc.threshold}{Minimum log fold change between the 2 populations. Default 0.25.}

\item{n}{Top number of genes to show based on fold change. Default 10.}
}
\value{
List of genes upregulated in in Pos and Neg cells
}
\description{
Calculates DEGs for Pos vs Neg cell populations and shows top upregulated genes for each population.
}
\author{
Kimberle Shen
}
