% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{positive_neighbor_histogram}
\alias{positive_neighbor_histogram}
\title{Frequency plot tabulating the number of positive neighbours each positive cell has.}
\usage{
positive_neighbor_histogram(obj, cell_list, k_geom = 15)
}
\arguments{
\item{obj}{Seurat v5 object with annotate_pos_neg ran}

\item{cell_list}{list of cellIDs in char format, eg c("65828963750382", "65816078848465", "65584150614472")}

\item{k_geom}{number of neighbours for each cell, default is 15}
}
\value{
Frequency plot tabulating the number of neighbours that are positive, for each cell, eg 0, 1, 2, 3
}
\description{
Frequency plot tabulating the number of positive neighbours each positive cell has.
}
\author{
Kimberle Shen
}
