% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{power_estimation}
\alias{power_estimation}
\title{Estimating power based on a gamma fit model in a scRNA-seq dataset to determine the ability to detect true DE, using scPower.}
\usage{
power_estimation(obj, perturbation, control, columnToUse = "gRNA", nSamples)
}
\arguments{
\item{obj}{Seurat v5 object with annotate_gRNA function ran}

\item{control}{Name of control cells in the gRNA column}

\item{columnToUse}{Name of the column where the perturbation and control labels are stored, default "gRNA"}

\item{pertubation}{Name of perturbation in the gRNA column}
}
\value{
Power estimation, between 0 to 1.
}
\description{
Estimating power based on a gamma fit model in a scRNA-seq dataset to determine the ability to detect true DE, using scPower.
}
\author{
Kimberle Shen
}
