% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{spatial_pos_neg}
\alias{spatial_pos_neg}
\title{Plot all cells on XY axes, with negative cells in black and positive cells in red.}
\usage{
spatial_pos_neg(obj)
}
\arguments{
\item{obj}{Seurat v5 object with annotate_pos_neg function ran}
}
\value{
Spatial plot of negative cells in black and positive cells in red.
}
\description{
Plot all cells on XY axes, with negative cells in black and positive cells in red.
}
\author{
Kimberle Shen
}
